Split PunctuationParser into two separate parsers
